requirements.txt
torch~=1.11.0+cu113
diffusers~=0.5.1
grpcio~=1.49.1